refactors shared resources between auto complete components
