+++ src/Autocomplete/Autocomplete.stories.mdx
@@ -8,7 +8,7 @@ import ButtonGroup from '../ButtonGroup';
 import Details from '../Details';
 import Label from '../Label';
 import Autocomplete, { RefDataAutocomplete } from './Autocomplete';
-import { Countries, CountriesAsRefData, CountriesByName } from './Countries-sb';
+import { Countries, CountriesAsRefData, CountriesByName } from '../storybook-resources/Countries-sb';
 
 <!-- Styles -->
 import './Autocomplete.stories.scss';
