+++ src/FormGroup/FormGroup.stories.mdx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Canvas, Meta, Props, Story } from '@storybook/addon-docs';
 import Autocomplete from '../Autocomplete/Autocomplete';
-import { CountriesByName } from '../Autocomplete/Countries-sb';
+import { CountriesByName } from '../storybook-resources/Countries-sb';
 import Details from '../Details';
 import FormGroup from './FormGroup';
 import Tag from '../Tag';
