+++ src/MultiSelectAutocomplete/MultiSelectAutocomplete.stories.mdx
@@ -9,7 +9,7 @@ import Details from '../Details';
 import Label from '../Label';
 import Tag from '../Tag';
 import MultiSelectAutocomplete from './MultiSelectAutocomplete';
-import { Countries, CountriesAsRefData } from './Countries-sb';
+import { Countries, MultiSelectCountriesAsRefData } from '../storybook-resources/Countries-sb';
 
 <!-- Styles -->
 import './MultiSelectAutocomplete.stories.scss';
@@ -37,7 +37,7 @@ In this example, the data items take the form
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -85,7 +85,6 @@ In this example, the data items take the form `{ "value": "#d4351c", "label": "R
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: false,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -122,7 +121,7 @@ In this example, the data items take the form `{ "value": "#d4351c", "label": "R
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -160,7 +159,7 @@ In this example, the data items take the form `{ "value": "#d4351c", "label": "R
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: true,
         readonly: false,
@@ -197,7 +196,7 @@ In this example, the data items take the form `{ "value": "#d4351c", "label": "R
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: true,
         readonly: false,
@@ -243,7 +242,7 @@ In this example, the data items take the form `{ "value": "#d4351c", "label": "R
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -289,8 +288,8 @@ Reference data items include `value` and `label` properties: `{ "value": "GB", "
       const ATTRS = {
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
-        options: CountriesAsRefData,
-        isMulti: true,
+        options: MultiSelectCountriesAsRefData,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -328,7 +327,7 @@ In this example, the placeholder text is `Select all that applies` for a multi s
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: true,
+        multi: true,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
@@ -366,7 +365,6 @@ In this example, the placeholder text is `Select an option` for a single select 
         id: DEFAULT_ID,
         fieldId: DEFAULT_ID,
         options: Countries,
-        isMulti: false,
         className: DEFAULT_CLASS,
         disabled: false,
         readonly: false,
