+++ src/storybook-resources/Countries-sb.test.js
@@ -1,9 +1,10 @@
-import { Countries, CountriesAsRefData, CountriesByName } from './Countries-sb';
+import { Countries, CountriesAsRefData, CountriesByName, MultiSelectCountriesAsRefData } from './Countries-sb';
 
 describe('Countries-sb', () => {
   it('should export arrays appropriately', () => {
     expect(Array.isArray(Countries)).toBeTruthy();
     expect(Array.isArray(CountriesAsRefData)).toBeTruthy();
     expect(Array.isArray(CountriesByName)).toBeTruthy();
+    expect(Array.isArray(MultiSelectCountriesAsRefData)).toBeTruthy();
   });
 });
